+++ python/GafferOSLTest/ShadingEngineTest.py
@@ -719,19 +719,25 @@ class ShadingEngineTest( GafferOSLTest.OSLTestCase ) :
 
 	def testTextureOrientation( self ) :
 
-		s = self.compileShader( pathlib.Path( __file__ ).parent / "shaders" / "uvTextureMap.osl" )
-		e = GafferOSL.ShadingEngine( IECoreScene.ShaderNetwork(
+		shader = self.compileShader( pathlib.Path( __file__ ).parent / "shaders" / "uvTextureMap.osl" )
+		shaderNetwork = IECoreScene.ShaderNetwork(
 			shaders = {
-				"output" : IECoreScene.Shader( s, "osl:surface", { "fileName" : ( pathlib.Path( __file__ ).parent / "images" / "vRamp.tx" ).as_posix() } )
+				"output" : IECoreScene.Shader( shader, "osl:surface", { "fileName" : ( pathlib.Path( __file__ ).parent / "images" / "vRamp.tx" ).as_posix() } )
 			},
 			output = "output"
-		) )
+		)
 
-		p = self.rectanglePoints()
-		r = e.shade( p )
+		for origin in ( GafferOSL.ShadingEngine.TextureOrigin.values.values() ) :
 
-		for i, c in enumerate( r["Ci"] ) :
-			self.assertAlmostEqual( c[1], p["v"][i], delta = 0.02 )
+			engine = GafferOSL.ShadingEngine( shaderNetwork, origin )
+			points = self.rectanglePoints()
+			result = engine.shade( points )
+
+			for i, c in enumerate( result["Ci"] ) :
+				if origin == GafferOSL.ShadingEngine.TextureOrigin.Bottom :
+					self.assertAlmostEqual( c[1], points["v"][i], delta = 0.02 )
+				else :
+					self.assertAlmostEqual( c[1], 1.0 - points["v"][i], delta = 0.02 )
 
 	def testDerivatives( self ) :